[BUGFIX] Undefined array key in install tool
Fixes two undefined array keys in the install system extension's
UpgradeWizardsService.php that cause errors in PHP 8.

Resolves: #95824
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72008
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Georg Ringer <georg.ringer@gmail.com>

+++ typo3/sysext/install/Classes/Updates/DatabaseRowsUpdateWizard.php
@@ -330,7 +330,7 @@ class DatabaseRowsUpdateWizard implements UpgradeWizardInterface, RepeatableInte
     protected function updateOrDeleteRow(Connection $connectionForTable, Connection $connectionForSysRegistry, string $table, int $uid, array $updatedFields, array $startPosition): void
     {
         $deleteField = $GLOBALS['TCA'][$table]['ctrl']['delete'] ?? null;
-        if ($deleteField === null && $updatedFields['deleted'] === 1) {
+        if ($deleteField === null && isset($updatedFields['deleted']) && $updatedFields['deleted'] === 1) {
             $connectionForTable->delete(
                 $table,
                 [
